Update Discord voice timeout
From my own testing, I've come to the conclusion that the default timeout is too short - the bot would stop playing music if the voice connection got lost for a moment or if it was moved to another channel. I think 5 seconds is plenty of time for it to reconnect to voice if it disconnects.

Thanks for this great project.